Revert "Dev ly"
