+++ pom.xml
@@ -12,7 +12,7 @@
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.5.4</version>
-		<relativePath/>
+		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>org.hnu</groupId>
 	<artifactId>precomputation</artifactId>
@@ -27,6 +27,7 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter</artifactId>
 		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
@@ -37,57 +38,6 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>com.baomidou</groupId>
-			<artifactId>mybatis-plus-boot-starter</artifactId>
-			<version>3.4.3</version>
-		</dependency>
-		<dependency>
-		<groupId>org.janusgraph</groupId>
-		<artifactId>janusgraph-driver</artifactId>
-		<version>0.6.2</version>
-	</dependency>
-	<dependency>
-		<groupId>org.apache.tinkerpop</groupId>
-		<artifactId>gremlin-driver</artifactId>
-		<version>3.5.3</version>
-	</dependency>
-		<!--
-		<dependency>
-			<groupId>org.janusgraph</groupId>
-			<artifactId>janusgraph-cql</artifactId>
-			<version>0.6.2</version>
-		</dependency>
-		-->
-
-		<dependency>
-			<groupId>org.hibernate</groupId>
-			<artifactId>hibernate-validator</artifactId>
-			<version>5.2.4.Final</version>
-		</dependency>
-		<dependency>
-			<groupId>org.projectlombok</groupId>
-			<artifactId>lombok</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.janusgraph</groupId>
-			<artifactId>janusgraph-hbase</artifactId>
-			<version>0.6.2</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.hbase</groupId>
-			<artifactId>hbase-client</artifactId>
-			<version>2.2.7</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.hadoop</groupId>
-			<artifactId>hadoop-hdfs</artifactId>
-			<version>2.8.5</version>
-		</dependency>
-
-
-
-
 	</dependencies>
 
 	<build>
