+++ precomputation-common/src/main/java/org/hnu/precomputation/common/view/dataset/DatasetAddParam.java
@@ -16,16 +16,4 @@ public class DatasetAddParam {
      * 数据集描述
      */
     private String description;
-    /**
-     * 属性
-     */
-    private String vertexProperty;
-
-    private String edgeProperty;
-
-    /**
-     * 任务名称
-     */
-    private String taskName;
-
 }
