--- precomputation-service/src/main/java/org/hnu/precomputation/service/Impl/Pair.java
@@ -1,6 +1,0 @@
-package org.hnu.precomputation.service.Impl;
-
-public class Pair {
-    public Long vertex1;
-    public Long vertex2;
-}
