--- precomputation-service/src/main/java/org/hnu/precomputation/service/dao/TaskMapper.java
@@ -1,10 +1,0 @@
-package org.hnu.precomputation.service.dao;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-import org.hnu.precomputation.common.model.dataset.Task;
-
-@Mapper
-public interface TaskMapper extends BaseMapper<Task>{
-
-}
