+++ precomputation-service/src/main/java/org/hnu/precomputation/service/service/DatasetService.java
@@ -7,8 +7,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 public class DatasetService {
     private final static Logger logger = LoggerFactory.getLogger("DatasetService");
@@ -27,13 +25,4 @@ public class DatasetService {
     public void update(Dataset dataset) {
         datasetDao.updateByPrimaryKey(dataset);
     }
-
-    public void delete(Long id) {
-        datasetDao.deleteByPrimaryKey(id);
-    }
-
-    public List<Dataset> list(){
-        return datasetDao.selectAll();
-    }
-
 }
