+++ precomputation-service/src/main/java/org/hnu/precomputation/service/service/NebulaGraphService.java
@@ -2,7 +2,6 @@ package org.hnu.precomputation.service.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 @Service
