--- precomputation-service/src/main/java/org/hnu/precomputation/service/service/TaskService.java
@@ -1,12 +1,0 @@
-package org.hnu.precomputation.service.service;
-
-import com.baomidou.mybatisplus.extension.service.IService;
-import org.hnu.precomputation.common.model.dataset.Task;
-import org.hnu.precomputation.service.dao.TaskMapper;
-import org.springframework.stereotype.Service;
-
-
-@Service
-public interface TaskService extends IService<Task> {
-
-}
