--- precomputation-service/src/main/resources/janusgraph-cql.properties
@@ -1,118 +1,0 @@
-# Copyright 2019 JanusGraph Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# JanusGraph configuration sample: Cassandra over a socket
-#
-# This file connects to a Cassandra daemon running on localhost via
-# CQL.  Cassandra must already be started before starting JanusGraph
-# with this file.
-
-# The implementation of graph factory that will be used by gremlin server
-#
-# Default:    org.janusgraph.core.JanusGraphFactory
-# Data Type:  String
-# Mutability: LOCAL
-gremlin.graph=org.janusgraph.core.JanusGraphFactory
-
-# The primary persistence provider used by JanusGraph.  This is required. 
-# It should be set one of JanusGraph's built-in shorthand names for its
-# standard storage backends (shorthands: berkeleyje, cql, hbase, inmemory)
-# or to the full package and classname of a custom/third-party
-# StoreManager implementation.
-#
-# Default:    (no default value)
-# Data Type:  String
-# Mutability: LOCAL
-storage.backend=cql
-
-# The hostname or comma-separated list of hostnames of storage backend
-# servers.  This is only applicable to some storage backends, such as
-# cassandra and hbase.
-#
-# Default:    127.0.0.1
-# Data Type:  class java.lang.String[]
-# Mutability: LOCAL
-storage.hostname=192.168.70.184
-
-# The name of JanusGraph's keyspace.  It will be created if it does not
-# exist.
-#
-# Default:    janusgraph
-# Data Type:  String
-# Mutability: LOCAL
-storage.cql.keyspace=zhijiang2
-
-# The name of the local or closest Cassandra datacenter. This value will
-# be passed into CqlSessionBuilder.withLocalDatacenter.
-#
-# Default:    datacenter1
-# Data Type:  String
-# Mutability: MASKABLE
-storage.cql.local-datacenter=datacenter1
-
-
-
-# Whether to enable JanusGraph's database-level cache, which is shared
-# across all transactions. Enabling this option speeds up traversals by
-# holding hot graph elements in memory, but also increases the likelihood
-# of reading stale data.  Disabling it forces each transaction to
-# independently fetch graph elements from storage before reading/writing
-# them.
-#
-# Default:    false
-# Data Type:  Boolean
-# Mutability: MASKABLE
-cache.db-cache = true
-
-# How long, in milliseconds, database-level cache will keep entries after
-# flushing them.  This option is only useful on distributed storage
-# backends that are capable of acknowledging writes without necessarily
-# making them immediately visible.
-#
-# Default:    50
-# Data Type:  Integer
-# Mutability: GLOBAL_OFFLINE
-#
-# Settings with mutability GLOBAL_OFFLINE are centrally managed in
-# JanusGraph's storage backend.  After starting the database for the first
-# time, this file's copy of this setting is ignored.  Use JanusGraph's
-# Management System to read or modify this value after bootstrapping.
-cache.db-cache-clean-wait = 20
-
-# Default expiration time, in milliseconds, for entries in the
-# database-level cache. Entries are evicted when they reach this age even
-# if the cache has room to spare. Set to 0 to disable expiration (cache
-# entries live forever or until memory pressure triggers eviction when set
-# to 0).
-#
-# Default:    10000
-# Data Type:  Long
-# Mutability: GLOBAL_OFFLINE
-#
-# Settings with mutability GLOBAL_OFFLINE are centrally managed in
-# JanusGraph's storage backend.  After starting the database for the first
-# time, this file's copy of this setting is ignored.  Use JanusGraph's
-# Management System to read or modify this value after bootstrapping.
-cache.db-cache-time = 180000
-
-# Size of JanusGraph's database level cache.  Values between 0 and 1 are
-# interpreted as a percentage of VM heap, while larger values are
-# interpreted as an absolute size in bytes.
-#
-# Default:    0.3
-# Data Type:  Double
-# Mutability: MASKABLE
-cache.db-cache-size = 0.5
-query.batch=true
-
