+++ precomputation-service/src/main/resources/mapper/dataset.xml
@@ -9,8 +9,6 @@
         <result column="name" jdbcType="VARCHAR" property="name"/>
         <result column="source" jdbcType="TINYINT" property="source"/>
         <result column="description" jdbcType="VARCHAR" property="description"/>
-        <result column="vertexProperty" jdbcType="VARCHAR" property="vertexProperty"/>
-        <result column="edgeProperty" jdbcType="VARCHAR" property="edgeProperty"/>
     </resultMap>
 
     <insert id="insert" parameterType="org.hnu.precomputation.common.model.dataset.Dataset"
@@ -26,12 +24,6 @@
             <if test="description != null">
                 description,
             </if>
-            <if test="vertexProperty!= null">
-                vertexProperty ,
-            </if>
-            <if test="edgeProperty!= null">
-                edgeProperty ,
-            </if>
         </trim>
         <trim prefix="values (" suffix=")" suffixOverrides=",">
             <if test="name != null">
@@ -43,12 +35,6 @@
             <if test="description != null">
                 #{description,jdbcType=VARCHAR},
             </if>
-            <if test="vertexProperty != null">
-                #{vertexProperty,jdbcType=VARCHAR},
-            </if>
-            <if test="edgeProperty != null">
-                #{edgeProperty,jdbcType=VARCHAR},
-            </if>
         </trim>
     </insert>
 
@@ -57,9 +43,9 @@
         from dataset
         where id = #{id}
     </select>
-
+    
     <select id="selectByPrimaryKey" resultMap="Dataset" parameterType="java.lang.Long">
-        select id, name, source, description, vertexProperty,edgeProperty
+        select id, name, source, description
         from dataset
         where id = #{id}
     </select>
@@ -76,15 +62,7 @@
             <if test="description != null">
                 description = #{description,jdbcType=VARCHAR},
             </if>
-            <if test="vertexProperty != null">
-                vertexProperty = #{vertexProperty,jdbcType=VARCHAR},
-            </if>
-            <if test="edgeProperty != null">
-                edgeProperty = #{edgeProperty,jdbcType=VARCHAR},
-            </if>
-
         </set>
         where id = #{id,jdbcType=BIGINT}
     </update>
-
 </mapper>
