+++ precomputation-web/src/main/java/org/hnu/precomputation/web/PrecomputationApplication.java
@@ -1,4 +1,4 @@
-package org.hnu.precomputation.service.web;
+package org.hnu.precomputation.web;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
