+++ precomputation-web/src/main/resources/application-dev.yml
@@ -1,32 +1,29 @@
 server:
   port: 8199
-
+     
 spring:
-  datasource:
-    type: com.alibaba.druid.pool.DruidDataSource
-    druid:
-      driver-class-name: com.mysql.cj.jdbc.Driver
-      # 这里替换上自己的mysql服务连接信息
-      url: jdbc:mysql://192.168.70.184:3306/zhijiang?useUnicode=true&characterEncoding=UTF-8&allowMultiQueries=true&serverTimezone=Asia/Shanghai
-      username: zhijiang
-      password: 123456
-      initial-size: 10
-      max-active: 20
-      max-wait: 5000
-      time-between-eviction-runs-millis: 30000
-      min-evictable-idle-time-millis: 100000
-      test-while-idle: true
-      test-on-borrow: true
-      test-on-return: false
-      pool-prepared-statements: true
-      max-pool-prepared-statement-per-connection-size: 10
+    datasource:
+        type: com.alibaba.druid.pool.DruidDataSource
+        druid:
+          driver-class-name: com.mysql.cj.jdbc.Driver
+          # 这里替换上自己的mysql服务连接信息
+          url: jdbc:mysql://10.5.24.98:3306/precomputation?useUnicode=true&characterEncoding=UTF-8&allowMultiQueries=true&serverTimezone=Asia/Shanghai
+          username: root
+          password: Zhejianglab@123
+          initial-size: 10
+          max-active: 20
+          max-wait: 5000
+          time-between-eviction-runs-millis: 30000
+          min-evictable-idle-time-millis: 100000
+          test-while-idle: true
+          test-on-borrow: true
+          test-on-return: false
+          pool-prepared-statements: true
+          max-pool-prepared-statement-per-connection-size: 10
 
 mybatis:
   mapper-locations: classpath*:mapper/*.xml
   type-aliases-package:
   configuration:
     map-underscore-to-camel-case: true
-
-
-
