+++ precomputation-web/src/main/resources/application.yml
@@ -8,4 +8,3 @@ spring:
       max-request-size: 2048MB
       max-file-size: 2048MB
 
-
